Login page, nav bar, suggestion page
In the first session, I added a really simple login page that utilizes a nav bar that (at least at this point) will live at the top of every page. Also set up basic routing to allow clicking the login button to take you to the dummy info on the suggestion page.